move to domain/models
