+++ src/com/apmath/loans/application/v1/models/PaymentInterface.kt
@@ -1,7 +1,7 @@
 package com.apmath.loans.application.v1.models
 
-import com.apmath.loans.domain.data.Currency
-import com.apmath.loans.domain.data.Money
+import com.apmath.loans.domain.models.data.Currency
+import com.apmath.loans.domain.models.data.Money
 
 interface PaymentInterface {
     val payment: Money
