--- src/com/apmath/loans/domain/connections/request/PaymentInterface.kt
@@ -1,5 +1,0 @@
-package com.apmath.loans.domain.connections.request
-
-interface PaymentInterface {
-    val date: String
-}
