--- src/com/apmath/loans/domain/data/Money.kt
@@ -1,3 +1,0 @@
-package com.apmath.loans.domain.data
-
-inline class Money(val value: Long)
