+++ src/com/apmath/loans/domain/models/connections/request/LoanInterface.kt
@@ -1,6 +1,6 @@
-package com.apmath.loans.domain.connections.request
+package com.apmath.loans.domain.models.connections.request
 
-import com.apmath.loans.domain.data.Money
+import com.apmath.loans.domain.models.data.Money
 
 interface LoanInterface {
     val interest: Int
