+++ src/com/apmath/loans/domain/models/connections/request/PaymentInterface.kt
@@ -1,0 +1,5 @@
+package com.apmath.loans.domain.models.connections.request
+
+interface PaymentInterface {
+    val date: String
+}
