+++ src/com/apmath/loans/domain/models/data/Money.kt
@@ -1,0 +1,3 @@
+package com.apmath.loans.domain.models.data
+
+inline class Money(val value: Long)
