+++ src/com/apmath/loans/domain/models/data/Type.kt
@@ -1,4 +1,4 @@
-package com.apmath.loans.domain.data
+package com.apmath.loans.domain.models.data
 
 enum class Type {
     REGULAR, EARLY, NEXT
